Improve path smoothing and culling
- Cull unnecessary nodes in the generated path by ray casting between
nodes to find the minimum nodes.
- Push generated pathing nodes away from walls so the bot doesn't try to
climb right against the wall.
- Improve wall avoidance steering behavior.
- Switch from Manhattan to Euclidean distance heuristic for AStar.
- Fix stuck detection for larger ships.

+++ screenbot/MemorySensor.cpp
@@ -115,7 +115,7 @@ void MemorySensor::DetectName() {
 void MemorySensor::DetectPosition() {
     const unsigned int PosStructID = 0x4AC7C0;
 
-    std::vector<uintptr_t> found = Memory::FindU32(m_ProcessHandle, PosStructID);
+    std::vector<unsigned int> found = Memory::FindU32(m_ProcessHandle, PosStructID);
     uintptr_t address = 0;
 
     if (found.size() == 1) {
